added checks

+++ src/services/userService.js
@@ -83,6 +83,13 @@ async function createUser(user){
      throw new Error("User is not found");
    }
 
+   if(user.login){
+    const u = await getUserByLogin(user.login);
+    if(u && u.login != user.login){
+        throw new Error("This login is already in use by another account");
+    }
+   }
+
    if(data.image)deleteFile(user.image);
    if(data.password)data.password = await getHesh(data.password);
